+++ app/app.go
@@ -170,7 +170,6 @@ func NewNibiruApp(
 	app.initModuleManager(encodingConfig, skipGenesisInvariants)
 
 	app.setupUpgrades()
-
 	// NOTE: Any module instantiated in the module manager that is later modified
 	// must be passed by reference here.