+++ app/upgrades/v1_0_2/constants.go
@@ -1,0 +1,23 @@
+package v1_0_2
+
+import (
+	"github.com/cosmos/cosmos-sdk/store/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/module"
+	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
+
+	"github.com/NibiruChain/nibiru/app/upgrades"
+)
+
+const UpgradeName = "v1.0.2"
+
+// pretty much a no-op store upgrade to test the upgrade process and include the newer version of rocksdb + cosmos-sdk
+var Upgrade = upgrades.Upgrade{
+	UpgradeName: UpgradeName,
+	CreateUpgradeHandler: func(mm *module.Manager, cfg module.Configurator) upgradetypes.UpgradeHandler {
+		return func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+			return mm.RunMigrations(ctx, cfg, fromVM)
+		}
+	},
+	StoreUpgrades: types.StoreUpgrades{},
+}